clean up the mem segment abstraction
